Added fix to handle mismatched cases
If the label that you want to manage exists on a given repository but the case doesn't match (e.g. 'tech debt' vs 'Tech Debt') then it would error. This adds a fix to handle that edge case cleanly and update the existing label with the specified case passed on the command line.